refactor: establish research domain family
