+++ docs/guide/agent-missions.md
@@ -508,6 +508,8 @@ archetype.app.missions
 └── agent_service.py   # composition and durable workflow boundary
 
 archetype.physical_ai  # physical state, policies, and external-step processors
+
+archetype.research     # research ledger state and pure runner decoding
 ```
 
 The cleanup direction is:
@@ -519,6 +521,7 @@ The cleanup direction is:
 | `archetype.experiments` | Remove; keep only code with a clear owner under missions, research, or physical AI. |
 | Trajectory helpers | Keep pure projections under missions; compose query- and evaluation-backed workflows in the app layer. |
 | Physical-AI prototypes | Moved into the registered `archetype.physical_ai` family; application rollout/evaluation workflows remain separate. |
+| Research ledger state | Moved into `archetype.research`; `archetype.app.research` retains world/simulation orchestration. |
 | Former `archetype.contrib` observability shim | Removed; retained vendor-neutral vocabulary lives in `archetype._obs`. |
 
 Those moves are not prerequisites for V1 and should not be smuggled into its
