+++ tests/app/test_autoresearch_ledger.py
@@ -20,9 +20,9 @@ from uuid_utils import UUID
 from archetype.app.container import ServiceContainer
 from archetype.app.models import EpisodeConfig
 from archetype.app.research.contracts import AutoResearchConfig, EvaluationResult
-from archetype.app.research.models import BranchHead, Experiment, Result, Run, RunStatus
 from archetype.core.component import Component
 from archetype.core.config import RunConfig, StorageConfig, WorldConfig
+from archetype.research import BranchHead, Experiment, Result, Run, RunStatus
 
 
 class Tag(Component):
